pyproject.toml
[project]
name = "key_generate_login"
version = "0.1.0"
description = ""
authors = [
    {name = "william ferrari",email = "willch3estudos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "mongoengine (>=0.29.1,<0.30.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
